requirements.txt
python-pptx
streamlit
